requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.4
beautifulsoup4==4.12.2
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
dataclasses-json==0.6.1
elastic-transport==8.4.1
elasticsearch==8.10.0
emoji==2.8.0
filetype==1.2.0
frozenlist==1.4.0
greenlet==3.0.0
idna==3.4
importlib-metadata==6.8.0
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kor==1.0.0
langchain==0.0.309
langdetect==1.0.9
langsmith==0.0.43
lxml==4.9.3
Markdown==3.4.4
markdownify==0.11.6
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.0
openai==0.28.1
packaging==23.2
pandas==1.5.3
pycodestyle==2.11.0
pydantic==2.4.2
pydantic_core==2.10.1
pyee==8.2.2
pyppeteer==1.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
tabulate==0.9.0
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
unstructured==0.10.19
urllib3==1.26.17
websockets==10.4
yarl==1.9.2
zipp==3.17.0
